Move issue-fix reviewer CLI into capability boundary
